admin flow added

+++ src/app/api/admins/category/route.ts
@@ -1,18 +1,36 @@
 import prisma from '@/db/prisma';
 import { NextRequest, NextResponse } from 'next/server';
 import { validateRequest } from '@/lib/validate';
-import { categoryReq, categorySchema } from '@/Schemas/categories';
+import {
+  categoryReq,
+  categorySchema,
+  deleteCategoryType,
+} from '@/Schemas/categories';
 
-export async function GET(request: NextRequest) {
+export async function GET() {
   try {
     const categories = await prisma.category.findMany({
       select: {
         id: true,
         name: true,
+        _count: {
+          select: { recipes: true },
+        },
       },
     });
+
+    const simplifiedResult = categories.map((category) => ({
+      id: category.id,
+      name: category.name,
+      count: category._count.recipes, // Extract the number
+    }));
+
     return NextResponse.json(
-      { success: true, message: `all listed categories`, data: categories },
+      {
+        success: true,
+        message: `all listed categories`,
+        data: simplifiedResult,
+      },
       { status: 201 }
     );
   } catch (error) {
@@ -37,8 +55,12 @@ export async function POST(request: NextRequest) {
       data: {
         name: body.name,
       },
+      select: {
+        id: true,
+        name: true,
+      },
     });
-    console.log('cat created:', newCategory);
+
     return NextResponse.json(
       { success: true, message: `New category Created` },
       { status: 201 }
@@ -56,10 +78,25 @@ export async function DELETE(request: NextRequest) {
   try {
     const body = await request.json();
 
-    const validate = validateRequest(categoryReq, body);
+    const validate = validateRequest(deleteCategoryType, body);
+
     if (validate instanceof NextResponse) {
       return validate;
     }
+    //check first does id exits or not also take care of related table id's
+    const checkItem = await prisma.category.findFirst({
+      where: {
+        ...body,
+      },
+    });
+
+    if (!checkItem) {
+      return NextResponse.json(
+        { success: false, message: "Coundn't find the recipe!!" },
+        { status: 500 }
+      );
+    }
+
     await prisma.category.delete({
       where: {
         ...body,
@@ -87,6 +124,20 @@ export async function PATCH(request: NextRequest) {
       return validate;
     }
 
+    //addition check if id exists or not
+    const checkItem = await prisma.category.findFirst({
+      where: {
+        ...body,
+      },
+    });
+
+    if (!checkItem) {
+      return NextResponse.json(
+        { success: false, message: "Coundn't find the recipe!!" },
+        { status: 500 }
+      );
+    }
+
     const updatedCategory = await prisma.category.update({
       where: {
         id: body.id,
@@ -96,13 +147,11 @@ export async function PATCH(request: NextRequest) {
       },
     });
 
-    console.log(updatedCategory);
     return NextResponse.json(
       { success: true, message: `Category name updated!` },
       { status: 200 }
     );
   } catch (error) {
-    console.log(error);
     return NextResponse.json(
       { success: false, message: 'Internal Server Error!' },
       { status: 500 }
